+++ source/extensions/isaacsim.replicator.examples/python/tests/test_sdg_useful_snippets_timeline_based.py
@@ -370,4 +370,8 @@ class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):
             omni.usd.get_context().get_rendering_event_stream().create_subscription_to_pop(on_stage_render_event)
         )
         app_update_events = []
-        app_sub = omni.kit.app.get_app().get_update_event_stream().create_subscription_to_pop(on_app_update)
+        app_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
+            event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
+            on_event=on_app_update,
+            observer_name="TestSDGUsefulSnippets.test_sdg_snippet_subscribers_and_events.on_app_update",
+        )