+++ source/extensions/isaacsim.core.api/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [4.2.22] - 2025-03-24
+### Changed
+- Migrate to Events 2.0
+
 ## [4.2.21] - 2025-03-11
 ### Changed
 - Switch asset root for tests to internal nucleus